+++ packages/common/dictionaryToObject.js
@@ -1,0 +1,4 @@
+import set from 'lodash/set';
+
+export default dictionary =>
+    dictionary.reduce((acc, entry) => set(acc, entry.key, entry.value), {});
