+++ packages/common/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "common",
+  "private": true,
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "jest"
+  },
+  "dependencies": {
+    "lodash": "4.17.15"
+  },
+  "devDependencies": {
+    "@babel/core": "7.4.5",
+    "@babel/preset-env": "7.4.5",
+    "jest": "24.7.1"
+  },
+  "author": "",
+  "license": "ISC"
+}
