+++ packages/polyglot-po-loader/package.json
@@ -4,9 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
+    "common": "^1.0.0",
     "lodash": "4.17.11",
     "pofile": "1.1.0"
   },
